feat/entity: Switch propriety of Dishes and Categories in Database
